--- server/module-api/src/test/java/com/momo/common/SpringContainerTest.java
@@ -1,17 +1,0 @@
-package com.momo.common;
-
-import org.junit.jupiter.api.AfterEach;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class SpringContainerTest {
-
-    @Autowired
-    private DatabaseCleaner databaseCleaner;
-
-    @AfterEach
-    void tearDown() {
-        databaseCleaner.cleanUp();
-    }
-}
