+++ server/module-core/src/test/java/com/momo/domain/group/entity/GroupListenerTest.java
@@ -1,14 +1,12 @@
-package com.momo.group;
+package com.momo.domain.group.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.momo.common.SpringContainerTest;
+import com.momo.common.RepositoryTest;
 import com.momo.domain.achievementrate.entity.GroupAchievementRate;
 import com.momo.domain.achievementrate.repository.GroupAchievementRateRepository;
-import com.momo.domain.group.entity.Group;
 import com.momo.domain.group.repository.GroupRepository;
 import com.momo.domain.user.entity.User;
-import com.momo.domain.user.repository.UserRepository;
 import java.util.List;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -17,13 +15,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 @DisplayName("모임 리스너 테스트")
-public class GroupListenerTest extends SpringContainerTest {
+public class GroupListenerTest extends RepositoryTest {
 
     @Autowired
     private GroupAchievementRateRepository groupAchievementRateRepository;
-
-    @Autowired
-    private UserRepository userRepository;
 
     @Autowired
     private GroupRepository groupRepository;
@@ -32,13 +27,13 @@ public class GroupListenerTest extends SpringContainerTest {
 
     @BeforeEach
     void before() {
-        User user = userRepository.save(
+        User user = save(
             User.builder()
                 .providerId("test")
                 .nickname("유저")
                 .build()
         );
-        group = groupRepository.save(
+        group = save(
             Group.builder()
                 .name("모임 이름")
                 .manager(user)
@@ -49,7 +44,7 @@ public class GroupListenerTest extends SpringContainerTest {
     @Test
     void 모임을_저장하면_저장_리스너가_실행된다() {
         verifyGroupPersistListener();
-        verifyExistsGroupAchievementRate();
+        verifyExistsGroupAchievementRateField();
     }
 
     private void verifyGroupPersistListener() {
@@ -63,7 +58,7 @@ public class GroupListenerTest extends SpringContainerTest {
         );
     }
 
-    private void verifyExistsGroupAchievementRate() {
+    private void verifyExistsGroupAchievementRateField() {
         Group actual = groupRepository.findById(this.group.getId()).get();
         assertThat(actual.getAchievementRate()).isNotNull();
     }
