+++ server/module-core/src/test/java/com/momo/domain/group/entity/ParticipantListenerTest.java
@@ -1,16 +1,11 @@
-package com.momo.group;
+package com.momo.domain.group.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.momo.common.SpringContainerTest;
+import com.momo.common.RepositoryTest;
 import com.momo.domain.achievementrate.entity.ParticipantAchievementRate;
 import com.momo.domain.achievementrate.repository.ParticipantAchievementRateRepository;
-import com.momo.domain.group.entity.Group;
-import com.momo.domain.group.entity.Participant;
-import com.momo.domain.group.repository.GroupRepository;
-import com.momo.domain.group.repository.ParticipantRepository;
 import com.momo.domain.user.entity.User;
-import com.momo.domain.user.repository.UserRepository;
 import java.util.List;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -19,37 +14,28 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 @DisplayName("참여자 리스너 테스트")
-public class ParticipantListenerTest extends SpringContainerTest {
+public class ParticipantListenerTest extends RepositoryTest {
 
     @Autowired
     private ParticipantAchievementRateRepository participantAchievementRateRepository;
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Autowired
-    private GroupRepository groupRepository;
-
-    @Autowired
-    private ParticipantRepository participantRepository;
 
     private Participant participant;
 
     @BeforeEach
     void before() {
-        User user = userRepository.save(
+        User user = save(
             User.builder()
                 .providerId("test")
                 .nickname("유저")
                 .build()
         );
-        Group group = groupRepository.save(
+        Group group = save(
             Group.builder()
                 .name("모임 이름")
                 .manager(user)
                 .build()
         );
-        participant = participantRepository.save(
+        participant = save(
             Participant.builder()
                 .group(group)
                 .user(user)
